@@ -16,75 +16,103 @@
     "lib/"
   ],
   "peerDependencies": {
-    "pnpm-logger": "^0.5.4"
+    "@pnpm/logger": "^1.0.0"
   },
   "dependencies": {
+    "@pnpm/check-package": "^1.0.0",
+    "@pnpm/fs-locker": "^1.0.0",
+    "@pnpm/package-requester": "^0.6.0",
+    "@pnpm/pkgid-to-filename": "^1.0.0",
+    "@pnpm/types": "^1.7.0",
     "@types/byline": "^4.2.31",
     "@types/common-tags": "^1.2.5",
     "@types/load-json-file": "^2.0.5",
-    "@types/mz": "^0.0.31",
-    "@types/node": "^7.0.27",
+    "@types/mz": "^0.0.32",
+    "@types/node": "^9.3.0",
     "@types/nopt": "^3.0.29",
     "@types/npm": "^2.0.28",
-    "@types/ramda": "^0.0.15",
+    "@types/p-series": "^1.0.0",
+    "@types/ramda": "^0.25.8",
     "@types/rc": "^0.0.1",
     "@types/semver": "^5.3.30",
     "@types/update-notifier": "^1.0.0",
-    "@zkochan/cmd-shim": "^2.2.3",
+    "@zkochan/cmd-shim": "^2.2.4",
+    "@zkochan/npm-package-arg": "^1.0.0",
     "arr-flatten": "^1.0.1",
     "byline": "^5.0.0",
     "common-tags": "^1.4.0",
     "cross-spawn": "^5.0.0",
-    "dependency-path": "^1.1.0",
+    "dependency-path": "^1.2.0",
     "encode-registry": "^1.0.0",
-    "execa": "^0.7.0",
-    "find-up": "^2.1.0",
+    "execa": "^0.8.0",
     "get-npm-tarball-url": "^2.0.0",
     "graceful-fs": "^4.1.11",
     "is-ci": "^1.0.10",
     "is-inner-link": "^2.0.0",
     "is-subdir": "^1.0.1",
     "is-windows": "^1.0.0",
-    "load-json-file": "^2.0.0",
+    "load-json-file": "^4.0.0",
     "load-yaml-file": "^0.1.0",
     "mkdirp-promise": "^5.0.1",
-    "mz": "^2.6.0",
-    "ncp": "^2.0.0",
-    "node-gyp": "^3.6.0",
+    "mz": "^2.7.0",
     "normalize-path": "^2.1.1",
     "normalize-registry-url": "^1.0.0",
-    "npm": "^5.0.2",
-    "npm-package-arg": "^5.0.0",
-    "npm-registry-client": "^8.1.0",
+    "npm-lifecycle": "^2.0.0",
     "os-homedir": "^1.0.2",
     "p-filter": "^1.0.0",
     "p-limit": "^1.1.0",
-    "package-store": "^0.2.2",
+    "p-series": "^1.0.0",
+    "package-store": "^0.14.0",
+    "path-absolute": "^1.0.0",
     "path-exists": "^3.0.0",
     "path-name": "^1.0.0",
     "pnpm-install-checks": "^1.1.0",
-    "pnpm-shrinkwrap": "^3.1.0",
-    "proper-lockfile": "^2.0.0",
-    "ramda": "^0.24.1",
+    "pnpm-shrinkwrap": "^5.1.0",
+    "ramda": "^0.25.0",
     "read-package-json": "^2.0.5",
     "remove-all-except-outer-links": "^1.0.0",
+    "replace-string": "^1.1.0",
     "rimraf-then": "^1.0.1",
     "semver": "^5.3.0",
     "symlink-dir": "^1.1.0",
     "thenify": "^3.3.0",
+    "validate-npm-package-name": "^3.0.0",
+    "version-selector-type": "^2.0.0",
     "write-pkg": "^3.1.0",
     "write-yaml-file": "^1.0.0"
   },
   "devDependencies": {
-    "commitizen": "^2.8.6",
-    "cz-conventional-changelog": "^2.0.0",
-    "husky": "^0.13.2",
+    "@commitlint/cli": "^4.2.0",
+    "@commitlint/config-angular": "^4.2.0",
+    "@commitlint/prompt-cli": "^4.2.0",
+    "@pnpm/default-fetcher": "^0.3.1",
+    "@pnpm/default-resolver": "^0.1.1",
+    "@pnpm/logger": "^1.0.0",
+    "@types/mkdirp": "^0.5.1",
+    "@types/sinon": "^4.1.2",
+    "@types/tape": "^4.2.30",
+    "@zkochan/husky": "^0.0.0",
+    "anonymous-npm-registry-client": "^0.1.2",
+    "caw": "^2.0.0",
+    "deep-require-cwd": "^1.0.0",
+    "exists-link": "^2.0.0",
+    "isexe": "^2.0.0",
+    "mkdirp": "^0.5.1",
+    "ncp": "^2.0.0",
+    "npm-run-all": "^4.0.1",
     "npm-scripts-info": "^0.3.6",
-    "pnpm-logger": "^0.5.4",
-    "rimraf": "^2.5.4",
+    "package-preview": "^1.0.0",
+    "pnpm-registry-mock": "^1.17.0",
+    "read-pkg": "^3.0.0",
+    "rimraf": "^2.6.2",
+    "sepia": "^2.0.2",
+    "sinon": "^4.1.3",
+    "tape": "^4.8.0",
+    "tape-promise": "^2.0.1",
+    "ts-node": "^4.0.1",
     "tslint": "^5.4.2",
-    "typescript": "^2.4.1",
-    "validate-commit-msg": "^2.7.0"
+    "typescript": "^2.6.1",
+    "write-json-file": "^2.2.0"
   },
   "directories": {
     "test": "test"
