@@ -1,48 +1,5 @@
-import logger from 'pnpm-logger'
-import path = require('path')
-import lockfile = require('proper-lockfile')
-import mkdirp = require('mkdirp-promise')
-import crypto = require('crypto')
-import expandTilde from '../fs/expandTilde'
-
-async function delay(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms))
-}
-
-async function lock(
-  lockFilename: string,
-  opts: {firstTime: boolean, stale: number}
-): Promise<{}> {
-  const promise = new Promise((resolve, reject) => {
-    lockfile.lock(
-      lockFilename,
-      {realpath: false, stale: opts.stale},
-      async (err: Error & {code: string}) => {
-        if (err && err.code === 'ELOCKED') {
-          if (opts.firstTime) {
-            logger.warn('waiting for another installation to complete...')
-          }
-          await delay(200)
-          await lock(lockFilename, {firstTime: false, stale: opts.stale})
-          resolve()
-        } else if (err) {
-          reject(err)
-        } else {
-          resolve()
-        }
-      })
-  })
-  return promise as Promise<{}>
-}
-
-async function unlock(lockFilename: string): Promise<{}> {
-  const promise = new Promise(resolve =>
-    lockfile.unlock(
-      lockFilename,
-      {realpath: false},
-      resolve))
-  return promise as Promise<{}>
-}
+import logger from '@pnpm/logger'
+import lock from '@pnpm/fs-locker'
 
 export default async function withLock<T> (
   dir: string,
@@ -52,19 +9,19 @@ export default async function withLock<T> (
     locks: string,
   }
 ): Promise<T> {
-  dir = path.resolve(dir)
-  // TODO: expand tilde only once. Currently it can be done several times during installation
-  // because several locks might get created
-  const locksDir = expandTilde(opts.locks)
-  await mkdirp(locksDir)
-  const lockFilename = path.join(locksDir, crypto.createHash('sha1').update(dir).digest('hex'))
-  await lock(lockFilename, {firstTime: true, stale: opts.stale})
+  const unlock = await lock(dir, {
+    stale: opts.stale,
+    locks: opts.locks,
+    whenLocked () {
+      logger.warn('waiting for another installation to complete...')
+    }
+  })
   try {
     const result = await fn()
-    await unlock(lockFilename)
+    await unlock()
     return result
   } catch (err) {
-    await unlock(lockFilename)
+    await unlock()
     throw err;
   }
 }
