@@ -1,30 +1,31 @@
 import path = require('path')
 import pFilter = require('p-filter')
-import {PnpmOptions} from '../types'
-import extendOptions from './extendOptions'
+import extendOptions, {
+  StoreStatusOptions,
+} from './extendStoreStatusOptions'
 import getContext from './getContext'
-import {pkgIsUntouched as untouched} from 'package-store'
+import checkPackage from '@pnpm/check-package'
 import * as dp from 'dependency-path'
-import {streamParser} from 'pnpm-logger'
+import {streamParser} from '@pnpm/logger'
 
-export default async function (maybeOpts: PnpmOptions) {
+export default async function (maybeOpts: StoreStatusOptions) {
   const reporter = maybeOpts && maybeOpts.reporter
   if (reporter) {
     streamParser.on('data', reporter)
   }
-  const opts = extendOptions(maybeOpts)
+  const opts = await extendOptions(maybeOpts)
   const ctx = await getContext(opts)
-  if (!ctx.shrinkwrap) return []
+  if (!ctx.wantedShrinkwrap) return []
 
-  const pkgPaths = Object.keys(ctx.shrinkwrap.packages || {})
+  const pkgPaths = Object.keys(ctx.wantedShrinkwrap.packages || {})
     .map(id => {
       if (id === '/') return null
-      return dp.resolve(ctx.shrinkwrap.registry, id)
+      return dp.resolve(ctx.wantedShrinkwrap.registry, id)
     })
     .filter(pkgId => pkgId && !ctx.skipped.has(pkgId))
     .map((pkgPath: string) => path.join(ctx.storePath, pkgPath))
 
-  const modified = await pFilter(pkgPaths, async (pkgPath: string) => !await untouched(path.join(pkgPath, 'package')))
+  const modified = await pFilter(pkgPaths, async (pkgPath: string) => !await checkPackage(path.join(pkgPath, 'package')))
 
   if (reporter) {
     streamParser.removeListener('data', reporter)
