@@ -1,37 +1,58 @@
+import pkgIdToFilename from '@pnpm/pkgid-to-filename'
 import {
   Resolution,
-  PackageContentInfo,
-  pkgIdToFilename,
-} from 'package-store'
-import {Dependencies, Package} from '../types'
+  PackageFilesResponse,
+} from '@pnpm/package-requester'
+import {Dependencies} from '@pnpm/types'
 import R = require('ramda')
 import semver = require('semver')
-import logger from 'pnpm-logger'
+import logger from '@pnpm/logger'
+import {PackageManifest} from '@pnpm/types'
 import path = require('path')
+import {oneLine} from 'common-tags'
+import crypto = require('crypto')
 import {InstalledPackage} from '../install/installMultiple'
 import {TreeNode, TreeNodeMap} from '../api/install'
+import {
+  splitNodeId,
+  createNodeId,
+  ROOT_NODE_ID,
+} from '../nodeIdUtils'
 
 export type DependencyTreeNode = {
   name: string,
   // at this point the version is really needed only for logging
   version: string,
   hasBundledDependencies: boolean,
-  path: string,
+  centralLocation: string,
   modules: string,
-  fetchingFiles: Promise<PackageContentInfo>,
+  fetchingFiles: Promise<PackageFilesResponse>,
   resolution: Resolution,
-  hardlinkedLocation: string,
-  children: string[],
+  peripheralLocation: string,
+  children: {[alias: string]: string},
   // an independent package is a package that
   // has neither regular nor peer dependencies
   independent: boolean,
   optionalDependencies: Set<string>,
   depth: number,
   absolutePath: string,
+  prod: boolean,
   dev: boolean,
   optional: boolean,
   id: string,
   installable: boolean,
+  additionalInfo: {
+    deprecated?: string,
+    peerDependencies?: Dependencies,
+    bundleDependencies?: string[],
+    bundledDependencies?: string[],
+    engines?: {
+      node?: string,
+      npm?: string,
+    },
+    cpu?: string[],
+    os?: string[],
+  },
 }
 
 export type DependencyTreeNodeMap = {
@@ -43,169 +64,199 @@ export type DependencyTreeNodeMap = {
 
 export default function (
   tree: TreeNodeMap,
-  rootNodeIds: string[],
-  topPkgIds: string[],
+  rootNodeIdsByAlias: {[alias: string]: string},
   // only the top dependencies that were already installed
   // to avoid warnings about unresolved peer dependencies
   topParents: {name: string, version: string}[],
   independentLeaves: boolean,
   nodeModules: string
-): DependencyTreeNodeMap {
-  const pkgsByName = R.fromPairs(
-    topParents.map((parent: {name: string, version: string}): R.KeyValuePair<string, ParentRef> => [
-      parent.name,
-      {
-        version: parent.version,
-        depth: 0
-      }
-    ])
+): {
+  resolvedTree: DependencyTreeNodeMap,
+  rootAbsolutePathsByAlias: {[alias: string]: string},
+} {
+  const pkgsByName = Object.assign(
+    R.fromPairs(
+      topParents.map((parent: {name: string, version: string}): R.KeyValuePair<string, ParentRef> => [
+        parent.name,
+        {
+          version: parent.version,
+          depth: 0
+        }
+      ])
+    ),
+    toPkgByName(R.keys(rootNodeIdsByAlias).map(alias => ({alias, nodeId: rootNodeIdsByAlias[alias], node: tree[rootNodeIdsByAlias[alias]]})))
   )
 
-  const nodeIdToResolvedId = {}
+  const absolutePathsByNodeId = {}
   const resolvedTree: DependencyTreeNodeMap = {}
-  resolvePeersOfChildren(new Set(rootNodeIds), pkgsByName, {
+  resolvePeersOfChildren(rootNodeIdsByAlias, pkgsByName, {
     tree,
-    nodeIdToResolvedId,
+    absolutePathsByNodeId,
     resolvedTree,
     independentLeaves,
     nodeModules,
+    purePkgs: new Set(),
   })
 
   R.values(resolvedTree).forEach(node => {
-    node.children = node.children.map(child => nodeIdToResolvedId[child])
+    node.children = R.keys(node.children).reduce((acc, alias) => {
+      acc[alias] = absolutePathsByNodeId[node.children[alias]]
+      return acc
+    }, {})
   })
-  return resolvedTree
+  return {
+    resolvedTree,
+    rootAbsolutePathsByAlias: R.keys(rootNodeIdsByAlias).reduce((rootAbsolutePathsByAlias, alias) => {
+      rootAbsolutePathsByAlias[alias] = absolutePathsByNodeId[rootNodeIdsByAlias[alias]]
+      return rootAbsolutePathsByAlias
+    }, {})
+  }
 }
 
 function resolvePeersOfNode (
   nodeId: string,
-  parentPkgs: ParentRefs,
+  parentParentPkgs: ParentRefs,
   ctx: {
     tree: TreeNodeMap,
-    nodeIdToResolvedId: {[nodeId: string]: string},
+    absolutePathsByNodeId: {[nodeId: string]: string},
     resolvedTree: DependencyTreeNodeMap,
     independentLeaves: boolean,
     nodeModules: string,
+    purePkgs: Set<string>, // pure packages are those that don't rely on externally resolved peers
   }
-): Set<string> {
+): {[alias: string]: string} {
   const node = ctx.tree[nodeId]
+  if (ctx.purePkgs.has(node.pkg.id) && ctx.resolvedTree[node.pkg.id].depth <= node.depth) {
+    ctx.absolutePathsByNodeId[nodeId] = node.pkg.id
+    return {}
+  }
 
-  const childrenSet = new Set(node.children)
-  const unknownResolvedPeersOfChildren = resolvePeersOfChildren(childrenSet, parentPkgs, ctx)
+  const children = typeof node.children === 'function' ? node.children() : node.children
+  const parentPkgs = R.isEmpty(children)
+    ? parentParentPkgs
+    : {
+        ...parentParentPkgs,
+        ...toPkgByName(R.keys(children).map(alias => ({alias, nodeId: children[alias], node: ctx.tree[children[alias]]})))
+      }
+  const unknownResolvedPeersOfChildren = resolvePeersOfChildren(children, parentPkgs, ctx, nodeId)
 
   const resolvedPeers = R.isEmpty(node.pkg.peerDependencies)
-    ? new Set<string>()
-    : resolvePeers(node, Object.assign({}, parentPkgs,
-      toPkgByName(R.props<TreeNode>(node.children, ctx.tree))
-    ), ctx.tree)
+    ? {}
+    : resolvePeers(nodeId, node, parentPkgs, ctx.tree)
 
-  unknownResolvedPeersOfChildren.delete(nodeId)
-
-  const allResolvedPeers = union(
-    unknownResolvedPeersOfChildren,
-    resolvedPeers)
+  const allResolvedPeers = Object.assign(unknownResolvedPeersOfChildren, resolvedPeers)
 
   let modules: string
   let absolutePath: string
   const localLocation = path.join(ctx.nodeModules, `.${pkgIdToFilename(node.pkg.id)}`)
-  if (!allResolvedPeers.size) {
+  if (R.isEmpty(allResolvedPeers)) {
     modules = path.join(localLocation, 'node_modules')
     absolutePath = node.pkg.id
+    if (R.isEmpty(node.pkg.peerDependencies)) {
+      ctx.purePkgs.add(node.pkg.id)
+    }
   } else {
-    const peersFolder = createPeersFolderName(R.props<TreeNode>(Array.from(allResolvedPeers), ctx.tree).map(node => node.pkg))
+    const peersFolder = createPeersFolderName(
+      R.keys(allResolvedPeers).map(alias => ({
+        name: alias,
+        version: ctx.tree[allResolvedPeers[alias]].pkg.version,
+      })))
     modules = path.join(localLocation, peersFolder, 'node_modules')
     absolutePath = `${node.pkg.id}/${peersFolder}`
   }
 
-  ctx.nodeIdToResolvedId[nodeId] = absolutePath
+  ctx.absolutePathsByNodeId[nodeId] = absolutePath
   if (!ctx.resolvedTree[absolutePath] || ctx.resolvedTree[absolutePath].depth > node.depth) {
-    const independent = ctx.independentLeaves && !node.children.length && R.isEmpty(node.pkg.peerDependencies)
-    const pathToUnpacked = path.join(node.pkg.path, 'node_modules', node.pkg.name)
-    const hardlinkedLocation = !independent
+    const independent = ctx.independentLeaves && R.isEmpty(node.children) && R.isEmpty(node.pkg.peerDependencies)
+    const centralLocation = path.join(node.pkg.path, 'node_modules', node.pkg.name)
+    const peripheralLocation = !independent
       ? path.join(modules, node.pkg.name)
-      : pathToUnpacked
+      : centralLocation
     ctx.resolvedTree[absolutePath] = {
       name: node.pkg.name,
       version: node.pkg.version,
       hasBundledDependencies: node.pkg.hasBundledDependencies,
       fetchingFiles: node.pkg.fetchingFiles,
       resolution: node.pkg.resolution,
-      path: pathToUnpacked,
+      centralLocation,
       modules,
-      hardlinkedLocation,
+      peripheralLocation,
       independent,
       optionalDependencies: node.pkg.optionalDependencies,
-      children: Array.from(union(childrenSet, resolvedPeers)),
+      children: Object.assign(children, resolvedPeers),
       depth: node.depth,
       absolutePath,
+      prod: node.pkg.prod,
       dev: node.pkg.dev,
       optional: node.pkg.optional,
       id: node.pkg.id,
       installable: node.installable,
+      additionalInfo: node.pkg.additionalInfo,
     }
   }
   return allResolvedPeers
 }
 
-function addMany<T>(a: Set<T>, b: Set<T>) {
-  for (const el of Array.from(b)) {
-    a.add(el)
-  }
-  return a
-}
-
-function union<T>(a: Set<T>, b: Set<T>) {
-  return new Set(Array.from(a).concat(Array.from(b)))
-}
-
-function difference<T>(a: Set<T>, b: Set<T>) {
-  return new Set(Array.from(a).filter(el => !b.has(el)))
-}
-
 function resolvePeersOfChildren (
-  children: Set<string>,
-  parentParentPkgs: ParentRefs,
+  children: {
+    [alias: string]: string,
+  },
+  parentPkgs: ParentRefs,
   ctx: {
     tree: {[nodeId: string]: TreeNode},
-    nodeIdToResolvedId: {[nodeId: string]: string},
+    absolutePathsByNodeId: {[nodeId: string]: string},
     resolvedTree: DependencyTreeNodeMap,
     independentLeaves: boolean,
     nodeModules: string,
-  }
-): Set<string> {
-  const childrenArray = Array.from(children)
-  let allResolvedPeers = new Set()
-  const parentPkgs = Object.assign({}, parentParentPkgs,
-    toPkgByName(R.props<TreeNode>(childrenArray, ctx.tree))
-  )
+    purePkgs: Set<string>,
+  },
+  exceptNodeId?: string,
+): {[alias: string]: string} {
+  let allResolvedPeers: {[alias: string]: string} = {}
 
-  for (const child of childrenArray) {
-    addMany(allResolvedPeers, resolvePeersOfNode(child, parentPkgs, ctx))
+  for (const childNodeId of R.values(children)) {
+    Object.assign(allResolvedPeers, resolvePeersOfNode(childNodeId, parentPkgs, ctx))
   }
 
-  const unknownResolvedPeersOfChildren = difference(allResolvedPeers, children)
+  const unknownResolvedPeersOfChildren = R.keys(allResolvedPeers)
+    .filter(alias => !children[alias] && allResolvedPeers[alias] !== exceptNodeId)
+    .reduce((unknownResolvedPeersOfChildren, peer) => {
+      unknownResolvedPeersOfChildren[peer] = allResolvedPeers[peer]
+      return unknownResolvedPeersOfChildren
+    }, {})
 
   return unknownResolvedPeersOfChildren
 }
 
 function resolvePeers (
+  nodeId: string,
   node: TreeNode,
   parentPkgs: ParentRefs,
   tree: TreeNodeMap
-): Set<string> {
-  const resolvedPeers = new Set<string>()
+): {
+  [alias: string]: string
+} {
+  const resolvedPeers: {[alias: string]: string} = {}
   for (const peerName in node.pkg.peerDependencies) {
     const peerVersionRange = node.pkg.peerDependencies[peerName]
 
     const resolved = parentPkgs[peerName]
 
     if (!resolved || resolved.nodeId && !tree[resolved.nodeId].installable) {
-      logger.warn(`${node.pkg.id} requires a peer of ${peerName}@${peerVersionRange} but none was installed.`)
+      const friendlyPath = nodeIdToFriendlyPath(nodeId, tree)
+      logger.warn(oneLine`
+        ${friendlyPath ? `${friendlyPath}: ` : ''}${packageFriendlyId(node.pkg)}
+        requires a peer of ${peerName}@${peerVersionRange} but none was installed.`
+      )
       continue
     }
 
     if (!semver.satisfies(resolved.version, peerVersionRange)) {
-      logger.warn(`${node.pkg.id} requires a peer of ${peerName}@${peerVersionRange} but version ${resolved.version} was installed.`)
+      const friendlyPath = nodeIdToFriendlyPath(nodeId, tree)
+      logger.warn(oneLine`
+        ${friendlyPath ? `${friendlyPath}: ` : ''}${packageFriendlyId(node.pkg)}
+        requires a peer of ${peerName}@${peerVersionRange} but version ${resolved.version} was installed.`
+      )
     }
 
     if (resolved.depth === 0 || resolved.depth === node.depth + 1) {
@@ -215,9 +266,20 @@ function resolvePeers (
       continue
     }
 
-    if (resolved && resolved.nodeId) resolvedPeers.add(resolved.nodeId)
+    if (resolved && resolved.nodeId) resolvedPeers[peerName] = resolved.nodeId
   }
   return resolvedPeers
+}
+
+function packageFriendlyId (pkg: {name: string, version: string}) {
+  return `${pkg.name}@${pkg.version}`
+}
+
+function nodeIdToFriendlyPath (nodeId: string, tree: TreeNodeMap) {
+  const parts = splitNodeId(nodeId).slice(2, -2)
+  return R.tail(R.scan((prevNodeId, pkgId) => createNodeId(prevNodeId, pkgId), ROOT_NODE_ID, parts))
+    .map(nodeId => tree[nodeId].pkg.name)
+    .join(' > ')
 }
 
 type ParentRefs = {
@@ -231,18 +293,29 @@ type ParentRef = {
   nodeId?: string,
 }
 
-function toPkgByName (nodes: TreeNode[]): ParentRefs {
+function toPkgByName (nodes: {alias: string, nodeId: string, node: TreeNode}[]): ParentRefs {
   const pkgsByName: ParentRefs = {}
   for (const node of nodes) {
-    pkgsByName[node.pkg.name] = {
-      version: node.pkg.version,
+    pkgsByName[node.alias] = {
+      version: node.node.pkg.version,
       nodeId: node.nodeId,
-      depth: node.depth,
+      depth: node.node.depth,
     }
   }
   return pkgsByName
 }
 
-function createPeersFolderName(peers: InstalledPackage[]) {
-  return peers.map(peer => `${peer.name.replace('/', '!')}@${peer.version}`).sort().join('+')
+function createPeersFolderName(peers: {name: string, version: string}[]) {
+  const folderName = peers.map(peer => `${peer.name.replace('/', '!')}@${peer.version}`).sort().join('+')
+
+  // We don't want the folder name to get too long.
+  // Otherwise, an ENAMETOOLONG error might happen.
+  // see: https://github.com/pnpm/pnpm/issues/977
+  //
+  // A bigger limit might be fine but the md5 hash will be 32 symbols,
+  // so for consistency's sake, we go with 32.
+  if (folderName.length > 32) {
+    return crypto.createHash('md5').update(folderName).digest('hex')
+  }
+  return folderName
 }
