@@ -8,24 +8,26 @@ import {
   prune as pruneShrinkwrap,
 } from 'pnpm-shrinkwrap'
 import {DependencyTreeNodeMap, DependencyTreeNode} from './resolvePeers'
-import {Resolution} from 'package-store'
+import {Resolution} from '@pnpm/package-requester'
 import R = require('ramda')
-import {Package} from '../types'
+import {PackageJson, Dependencies} from '@pnpm/types'
 
 export default function (
   pkgsToLink: DependencyTreeNodeMap,
   shrinkwrap: Shrinkwrap,
-  pkg: Package
+  pkg: PackageJson
 ): Shrinkwrap {
   shrinkwrap.packages = shrinkwrap.packages || {}
   for (const dependencyAbsolutePath of R.keys(pkgsToLink)) {
     const dependencyPath = dp.relative(shrinkwrap.registry, dependencyAbsolutePath)
     const result = R.partition(
-      (childResolvedId: string) => pkgsToLink[dependencyAbsolutePath].optionalDependencies.has(pkgsToLink[childResolvedId].name),
-      pkgsToLink[dependencyAbsolutePath].children
+      (child) => pkgsToLink[dependencyAbsolutePath].optionalDependencies.has(pkgsToLink[child.nodeId].name),
+      R.keys(pkgsToLink[dependencyAbsolutePath].children).map(alias => ({alias, nodeId: pkgsToLink[dependencyAbsolutePath].children[alias]}))
     )
-    shrinkwrap.packages[dependencyPath] = toShrDependency({
+    shrinkwrap.packages[dependencyPath] = toShrDependency(pkgsToLink[dependencyAbsolutePath].additionalInfo, {
       dependencyAbsolutePath,
+      name: pkgsToLink[dependencyAbsolutePath].name,
+      version: pkgsToLink[dependencyAbsolutePath].version,
       id: pkgsToLink[dependencyAbsolutePath].id,
       dependencyPath,
       resolution: pkgsToLink[dependencyAbsolutePath].resolution,
@@ -35,6 +37,7 @@ export default function (
       pkgsToLink,
       prevResolvedDeps: shrinkwrap.packages[dependencyPath] && shrinkwrap.packages[dependencyPath].dependencies || {},
       prevResolvedOptionalDeps: shrinkwrap.packages[dependencyPath] && shrinkwrap.packages[dependencyPath].optionalDependencies || {},
+      prod: pkgsToLink[dependencyAbsolutePath].prod,
       dev: pkgsToLink[dependencyAbsolutePath].dev,
       optional: pkgsToLink[dependencyAbsolutePath].optional,
     })
@@ -43,17 +46,32 @@ export default function (
 }
 
 function toShrDependency (
+  pkg: {
+    deprecated?: string,
+    peerDependencies?: Dependencies,
+    bundleDependencies?: string[],
+    bundledDependencies?: string[],
+    engines?: {
+      node?: string,
+      npm?: string,
+    },
+    cpu?: string[],
+    os?: string[],
+  },
   opts: {
     dependencyAbsolutePath: string,
+    name: string,
+    version: string,
     id: string,
     dependencyPath: string,
     resolution: Resolution,
     registry: string,
-    updatedDeps: string[],
-    updatedOptionalDeps: string[],
+    updatedDeps: {alias: string, nodeId: string}[],
+    updatedOptionalDeps: {alias: string, nodeId: string}[],
     pkgsToLink: DependencyTreeNodeMap,
     prevResolvedDeps: ResolvedDependencies,
     prevResolvedOptionalDeps: ResolvedDependencies,
+    prod: boolean,
     dev: boolean,
     optional: boolean,
   }
@@ -64,20 +82,53 @@ function toShrDependency (
   const result = {
     resolution: shrResolution
   }
+  if (dp.isAbsolute(opts.dependencyPath)) {
+    result['name'] = opts.name
+
+    // There is no guarantee that a non-npmjs.org-hosted package
+    // is going to have a version field
+    if (opts.version) {
+      result['version'] = opts.version
+    }
+  }
   if (!R.isEmpty(newResolvedDeps)) {
     result['dependencies'] = newResolvedDeps
   }
   if (!R.isEmpty(newResolvedOptionalDeps)) {
     result['optionalDependencies'] = newResolvedOptionalDeps
   }
-  if (opts.dev) {
+  if (opts.dev && !opts.prod) {
     result['dev'] = true
+  } else if (opts.prod && !opts.dev) {
+    result['dev'] = false
   }
   if (opts.optional) {
     result['optional'] = true
   }
   if (opts.dependencyAbsolutePath !== opts.id) {
     result['id'] = opts.id
+  }
+  if (pkg.peerDependencies) {
+    result['peerDependencies'] = pkg.peerDependencies
+  }
+  if (pkg.engines) {
+    for (let engine of R.keys(pkg.engines)) {
+      if (pkg.engines[engine] === '*') continue
+      result['engines'] = result['engines'] || {}
+      result['engines'][engine] = pkg.engines[engine]
+    }
+  }
+  if (pkg.cpu) {
+    result['cpu'] = pkg.cpu
+  }
+  if (pkg.os) {
+    result['os'] = pkg.os
+  }
+  if (pkg.bundledDependencies || pkg.bundleDependencies) {
+    result['bundledDependencies'] = pkg.bundledDependencies || pkg.bundleDependencies
+  }
+  if (pkg.deprecated) {
+    result['deprecated'] = pkg.deprecated
   }
   return result
 }
@@ -87,16 +138,24 @@ function toShrDependency (
 // the `depth` property defines how deep should dependencies be checked
 function updateResolvedDeps (
   prevResolvedDeps: ResolvedDependencies,
-  updatedDeps: string[],
+  updatedDeps: {alias: string, nodeId: string}[],
   registry: string,
   pkgsToLink: DependencyTreeNodeMap
 ) {
   const newResolvedDeps = R.fromPairs<string>(
-    R.props<DependencyTreeNode>(updatedDeps, pkgsToLink)
-      .map((dep): R.KeyValuePair<string, string> => [
-        dep.name,
-        absolutePathToRef(dep.absolutePath, dep.name, dep.resolution, registry)
-      ])
+    updatedDeps
+      .map((dep): R.KeyValuePair<string, string> => {
+        const pkgToLink = pkgsToLink[dep.nodeId]
+        return [
+          dep.alias,
+          absolutePathToRef(pkgToLink.absolutePath, {
+            alias: dep.alias,
+            realName: pkgToLink.name,
+            resolution: pkgToLink.resolution,
+            standardRegistry: registry,
+          })
+        ]
+      })
   )
   return R.merge(
     prevResolvedDeps,
@@ -109,19 +168,19 @@ function toShrResolution (
   resolution: Resolution,
   registry: string
 ): ShrinkwrapResolution {
-  if (dp.isAbsolute(dependencyPath) || resolution.type !== undefined || !resolution.integrity) {
-    return resolution
+  if (dp.isAbsolute(dependencyPath) || resolution.type !== undefined || !resolution['integrity']) {
+    return resolution as ShrinkwrapResolution
   }
   // This might be not the best solution to identify non-standard tarball URLs in the long run
   // but it at least solves the issues with npm Enterprise. See https://github.com/pnpm/pnpm/issues/867
-  if (!resolution.tarball.includes('/-/')) {
+  if (!resolution['tarball'].includes('/-/')) {
     return {
-      integrity: resolution.integrity,
-      tarball: relativeTarball(resolution.tarball, registry),
+      integrity: resolution['integrity'],
+      tarball: relativeTarball(resolution['tarball'], registry),
     }
   }
   return {
-    integrity: resolution.integrity,
+    integrity: resolution['integrity'],
   }
 }
 
