@@ -1,7 +1,7 @@
 import tape = require('tape')
 import promisifyTape from 'tape-promise'
 import readPkg = require('read-pkg')
-import {installPkgs} from '../../src'
+import {installPkgs} from 'supi'
 import {
   prepare,
   testDefaults,
@@ -11,7 +11,7 @@ const test = promisifyTape(tape)
 
 test('tarball from npm registry', async function (t) {
   const project = prepare(t)
-  await installPkgs(['http://registry.npmjs.org/is-array/-/is-array-1.0.1.tgz'], testDefaults())
+  await installPkgs(['http://registry.npmjs.org/is-array/-/is-array-1.0.1.tgz'], await testDefaults())
 
   const m = project.requireModule('is-array')
 
@@ -25,7 +25,7 @@ test('tarball from npm registry', async function (t) {
 
 test('tarball not from npm registry', async function (t) {
   const project = prepare(t)
-  await installPkgs(['https://github.com/hegemonic/taffydb/tarball/master'], testDefaults())
+  await installPkgs(['https://github.com/hegemonic/taffydb/tarball/master'], await testDefaults())
 
   const m = project.requireModule('taffydb')
 
@@ -36,7 +36,7 @@ test('tarball not from npm registry', async function (t) {
 
 test('tarballs from GitHub (is-negative)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['is-negative@https://github.com/kevva/is-negative/archive/1d7e288222b53a0cab90a331f1865220ec29560c.tar.gz'], testDefaults())
+  await installPkgs(['is-negative@https://github.com/kevva/is-negative/archive/1d7e288222b53a0cab90a331f1865220ec29560c.tar.gz'], await testDefaults())
 
   const m = project.requireModule('is-negative')
 
