@@ -5,13 +5,13 @@ import {
   prepare,
   testDefaults,
 } from '../utils'
-import {installPkgs} from '../../src'
+import {installPkgs} from 'supi'
 
 const test = promisifyTape(tape)
 
 test('install with --independent-leaves', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['rimraf@2.5.1'], testDefaults({independentLeaves: true}))
+  await installPkgs(['rimraf@2.5.1'], await testDefaults({independentLeaves: true}))
 
   const m = project.requireModule('rimraf')
   t.ok(typeof m === 'function', 'rimraf() is available')
@@ -20,10 +20,10 @@ test('install with --independent-leaves', async function (t: tape.Test) {
 
 test('--independent-leaves throws exception when executed on node_modules installed w/o the option', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['is-positive'], testDefaults({independentLeaves: false}))
+  await installPkgs(['is-positive'], await testDefaults({independentLeaves: false}))
 
   try {
-    await installPkgs(['is-negative'], testDefaults({independentLeaves: true}))
+    await installPkgs(['is-negative'], await testDefaults({independentLeaves: true}))
     t.fail('installation should have failed')
   } catch (err) {
     t.ok(err.message.indexOf('This node_modules was not installed with the --independent-leaves option.') === 0)
@@ -32,10 +32,10 @@ test('--independent-leaves throws exception when executed on node_modules instal
 
 test('--no-independent-leaves throws exception when executed on node_modules installed with --independent-leaves', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['is-positive'], testDefaults({independentLeaves: true}))
+  await installPkgs(['is-positive'], await testDefaults({independentLeaves: true}))
 
   try {
-    await installPkgs(['is-negative'], testDefaults({independentLeaves: false}))
+    await installPkgs(['is-negative'], await testDefaults({independentLeaves: false}))
     t.fail('installation should have failed')
   } catch (err) {
     t.ok(err.message.indexOf('This node_modules was installed with --independent-leaves option.') === 0)
@@ -45,7 +45,7 @@ test('--no-independent-leaves throws exception when executed on node_modules ins
 test('global installation with --independent-leaves', async function (t: tape.Test) {
   prepare(t)
   const globalPrefix = path.resolve('..', 'global')
-  const opts = testDefaults({global: true, prefix: globalPrefix, independentLeaves: true})
+  const opts = await testDefaults({global: true, prefix: globalPrefix, independentLeaves: true})
   await installPkgs(['is-positive'], opts)
 
   // there was an issue when subsequent installations were removing everything installed prior
