@@ -18,12 +18,18 @@ import {
 } from '../utils'
 import loadJsonFile = require('load-json-file')
 const basicPackageJson = loadJsonFile.sync(path.join(__dirname, '../utils/simple-package.json'))
-import {install, installPkgs, uninstall} from '../../src'
+import {install, installPkgs, uninstall} from 'supi'
 import exists = require('path-exists')
 import isWindows = require('is-windows')
 import deepRequireCwd = require('deep-require-cwd')
 import sinon = require('sinon')
-import {StageLog, RootLog, ProgressLog} from 'pnpm-logger'
+import {
+  StageLog,
+  RootLog,
+  ProgressLog,
+  PackageJsonLog,
+  StatsLog,
+} from 'supi'
 import writeJsonFile = require('write-json-file')
 
 const IS_WINDOWS = isWindows()
@@ -34,41 +40,78 @@ if (!caw() && !IS_WINDOWS) {
 
 test('small with dependencies (rimraf)', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['rimraf@2.5.1'], testDefaults())
+  await installPkgs(['rimraf@2.5.1'], await testDefaults())
 
   const m = project.requireModule('rimraf')
   t.ok(typeof m === 'function', 'rimraf() is available')
   await project.isExecutable('.bin/rimraf')
 })
 
+test('ignoring some files in the dependency', async (t: tape.Test) => {
+  const project = prepare(t)
+
+  const ignoreFile = (filename: string) => filename === 'readme.md'
+  await installPkgs(['is-positive@1.0.0'], await testDefaults({}, {}, {ignoreFile}))
+
+  t.ok(await exists(path.resolve('node_modules', 'is-positive', 'package.json')), 'package.json was not ignored')
+  t.notOk(await exists(path.resolve('node_modules', 'is-positive', 'readme.md')), 'readme.md was ignored')
+})
+
 test('no dependencies (lodash)', async (t: tape.Test) => {
   const project = prepare(t)
   const reporter = sinon.spy()
 
-  await installPkgs(['lodash@4.0.0'], testDefaults({reporter}))
+  await addDistTag('lodash', '4.1.0', 'latest')
 
+  await installPkgs(['lodash@4.0.0'], await testDefaults({reporter}))
+
+  t.ok(reporter.calledWithMatch(<PackageJsonLog>{
+    name: 'pnpm:package-json',
+    level: 'debug',
+    initial: {name: 'project', version: '0.0.0'},
+  }), 'initial package.json logged')
   t.ok(reporter.calledWithMatch(<StageLog>{
     name: 'pnpm:stage',
     level: 'debug',
     message: 'resolution_done',
   }), 'resolution stage done logged')
-  t.ok(reporter.calledWithMatch({
-    level: 'info',
-    message: 'Creating dependency tree',
-  }), 'informed about creating dependency tree')
-  t.ok(reporter.calledWithMatch({
-    level: 'info',
-    message: 'Adding 1 packages to node_modules',
-  }), 'informed about adding new packages to node_modules')
+  // Not logged for now
+  // t.ok(reporter.calledWithMatch({
+  //   level: 'info',
+  //   message: 'Creating dependency tree',
+  // }), 'informed about creating dependency tree')
+  t.ok(reporter.calledWithMatch(<StatsLog>{
+    name: 'pnpm:stats',
+    level: 'debug',
+    added: 1,
+  }), 'added stat')
+  t.ok(reporter.calledWithMatch(<StatsLog>{
+    name: 'pnpm:stats',
+    level: 'debug',
+    removed: 0,
+ }), 'removed stat')
   t.ok(reporter.calledWithMatch(<RootLog>{
     name: 'pnpm:root',
     level: 'info',
     added: {
       name: 'lodash',
+      realName: 'lodash',
       version: '4.0.0',
       dependencyType: 'prod',
+      latest: '4.1.0',
     },
   }), 'added to root')
+  t.ok(reporter.calledWithMatch(<PackageJsonLog>{
+    name: 'pnpm:package-json',
+    level: 'debug',
+    updated: {
+      name: 'project',
+      version: '0.0.0',
+      dependencies: {
+        lodash: '^4.0.0',
+      },
+    },
+  }), 'updated package.json logged')
 
   const m = project.requireModule('lodash')
   t.ok(typeof m === 'function', '_ is available')
@@ -77,7 +120,7 @@ test('no dependencies (lodash)', async (t: tape.Test) => {
 
 test('scoped modules without version spec (@rstacruz/tap-spec)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['@rstacruz/tap-spec'], testDefaults())
+  await installPkgs(['@rstacruz/tap-spec'], await testDefaults())
 
   const m = project.requireModule('@rstacruz/tap-spec')
   t.ok(typeof m === 'function', 'tap-spec is available')
@@ -86,7 +129,7 @@ test('scoped modules without version spec (@rstacruz/tap-spec)', async function 
 test('scoped package with custom registry', async function (t) {
   const project = prepare(t)
 
-  await installPkgs(['@scoped/peer'], testDefaults({
+  await installPkgs(['@scoped/peer'], await testDefaults({
     // setting an incorrect default registry URL
     registry: 'http://localhost:9999/',
     rawNpmConfig: {
@@ -106,7 +149,7 @@ test('modules without version spec, with custom tag config', async function (t) 
   await addDistTag('dep-of-pkg-with-1-dep', '100.1.0', 'latest')
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', tag)
 
-  await installPkgs(['dep-of-pkg-with-1-dep'], testDefaults({tag}))
+  await installPkgs(['dep-of-pkg-with-1-dep'], await testDefaults({tag}))
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.0.0')
 })
@@ -117,7 +160,7 @@ test('installing a package by specifying a specific dist-tag', async function (t
   await addDistTag('dep-of-pkg-with-1-dep', '100.1.0', 'latest')
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', 'beta')
 
-  await installPkgs(['dep-of-pkg-with-1-dep@beta'], testDefaults())
+  await installPkgs(['dep-of-pkg-with-1-dep@beta'], await testDefaults())
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.0.0')
 })
@@ -128,11 +171,11 @@ test('update a package when installing with a dist-tag', async function (t: tape
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', 'latest')
   await addDistTag('dep-of-pkg-with-1-dep', '100.1.0', 'beta')
 
-  await installPkgs(['dep-of-pkg-with-1-dep'], testDefaults({saveDev: true}))
+  await installPkgs(['dep-of-pkg-with-1-dep'], await testDefaults({saveDev: true}))
 
   const reporter = sinon.spy()
 
-  await installPkgs(['dep-of-pkg-with-1-dep@beta'], testDefaults({saveDev: true, reporter}))
+  await installPkgs(['dep-of-pkg-with-1-dep@beta'], await testDefaults({saveDev: true, reporter}))
 
   t.ok(reporter.calledWithMatch(<RootLog>{
     name: 'pnpm:root',
@@ -154,6 +197,7 @@ test('update a package when installing with a dist-tag', async function (t: tape
     },
   }), 'reported new version added to the root')
 
+  await project.has('dep-of-pkg-with-1-dep')
   await project.storeHas('dep-of-pkg-with-1-dep', '100.1.0')
 
   const pkg = await readPkg()
@@ -162,7 +206,7 @@ test('update a package when installing with a dist-tag', async function (t: tape
 
 test('scoped modules with versions (@rstacruz/tap-spec@4.1.1)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['@rstacruz/tap-spec@4.1.1'], testDefaults())
+  await installPkgs(['@rstacruz/tap-spec@4.1.1'], await testDefaults())
 
   const m = project.requireModule('@rstacruz/tap-spec')
   t.ok(typeof m === 'function', 'tap-spec is available')
@@ -170,7 +214,7 @@ test('scoped modules with versions (@rstacruz/tap-spec@4.1.1)', async function (
 
 test('scoped modules (@rstacruz/tap-spec@*)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['@rstacruz/tap-spec@*'], testDefaults())
+  await installPkgs(['@rstacruz/tap-spec@*'], await testDefaults())
 
   const m = project.requireModule('@rstacruz/tap-spec')
   t.ok(typeof m === 'function', 'tap-spec is available')
@@ -178,7 +222,7 @@ test('scoped modules (@rstacruz/tap-spec@*)', async function (t) {
 
 test('multiple scoped modules (@rstacruz/...)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['@rstacruz/tap-spec@*', '@rstacruz/travis-encrypt@*'], testDefaults())
+  await installPkgs(['@rstacruz/tap-spec@*', '@rstacruz/travis-encrypt@*'], await testDefaults())
 
   t.equal(typeof project.requireModule('@rstacruz/tap-spec'), 'function', 'tap-spec is available')
   t.equal(typeof project.requireModule('@rstacruz/travis-encrypt'), 'function', 'travis-encrypt is available')
@@ -186,7 +230,7 @@ test('multiple scoped modules (@rstacruz/...)', async function (t) {
 
 test('nested scoped modules (test-pnpm-issue219 -> @zkochan/test-pnpm-issue219)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['test-pnpm-issue219@1.0.2'], testDefaults())
+  await installPkgs(['test-pnpm-issue219@1.0.2'], await testDefaults())
 
   const m = project.requireModule('test-pnpm-issue219')
   t.ok(m === 'test-pnpm-issue219,@zkochan/test-pnpm-issue219', 'nested scoped package is available')
@@ -195,7 +239,7 @@ test('nested scoped modules (test-pnpm-issue219 -> @zkochan/test-pnpm-issue219)'
 test('idempotency (rimraf)', async (t: tape.Test) => {
   const project = prepare(t)
   const reporter = sinon.spy()
-  const opts = testDefaults({reporter})
+  const opts = await testDefaults({reporter})
 
   await installPkgs(['rimraf@2.5.1'], opts)
 
@@ -229,13 +273,13 @@ test('idempotency (rimraf)', async (t: tape.Test) => {
 
 test('reporting adding root package', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   await project.storeHas('flatten', '1.0.2')
 
   const reporter = sinon.spy()
 
-  await installPkgs(['flatten@1.0.2'], testDefaults({reporter}))
+  await installPkgs(['flatten@1.0.2'], await testDefaults({reporter}))
 
   t.ok(reporter.calledWithMatch(<RootLog>{
     name: 'pnpm:root',
@@ -248,13 +292,13 @@ test('reporting adding root package', async (t: tape.Test) => {
   }), 'reported that flatten added to the root')
 })
 
-test('overwriting (magic-hook@2.0.0 and @0.1.0)', async function (t) {
+test('overwriting (magic-hook@2.0.0 and @0.1.0)', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   await project.storeHas('flatten', '1.0.2')
 
-  await installPkgs(['magic-hook@0.1.0'], testDefaults())
+  await installPkgs(['magic-hook@0.1.0'], await testDefaults())
 
   // flatten is not removed from store even though it is unreferenced
   // store should be pruned to have this removed
@@ -266,23 +310,23 @@ test('overwriting (magic-hook@2.0.0 and @0.1.0)', async function (t) {
 
 test('overwriting (is-positive@3.0.0 with is-positive@latest)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['is-positive@3.0.0'], testDefaults({save: true}))
+  await installPkgs(['is-positive@3.0.0'], await testDefaults({save: true}))
 
   await project.storeHas('is-positive', '3.0.0')
 
-  await installPkgs(['is-positive@latest'], testDefaults({save: true}))
+  await installPkgs(['is-positive@latest'], await testDefaults({save: true}))
 
   await project.storeHas('is-positive', '3.1.0')
 })
 
 test('forcing', async function (t) {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPath = path.resolve('node_modules', 'magic-hook', 'dist')
   await rimraf(distPath)
 
-  await installPkgs(['magic-hook@2.0.0'], testDefaults({force: true}))
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults({force: true}))
 
   const distPathExists = await exists(distPath)
   t.ok(distPathExists, 'magic-hook@2.0.0 dist folder reinstalled')
@@ -290,12 +334,12 @@ test('forcing', async function (t) {
 
 test('argumentless forcing', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPath = path.resolve('node_modules', 'magic-hook', 'dist')
   await rimraf(distPath)
 
-  await install(testDefaults({force: true}))
+  await install(await testDefaults({force: true}))
 
   const distPathExists = await exists(distPath)
   t.ok(distPathExists, 'magic-hook@2.0.0 dist folder reinstalled')
@@ -303,12 +347,12 @@ test('argumentless forcing', async function (t: tape.Test) {
 
 test('no forcing', async function (t) {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPath = path.resolve('node_modules', 'magic-hook', 'dist')
   await rimraf(distPath)
 
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPathExists = await exists(distPath)
   t.notOk(distPathExists, 'magic-hook@2.0.0 dist folder not reinstalled')
@@ -316,14 +360,14 @@ test('no forcing', async function (t) {
 
 test('refetch package to store if it has been modified', async function (t) {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPathInStore = await project.resolve('magic-hook', '2.0.0', 'dist')
   await rimraf(distPathInStore)
   await rimraf('node_modules')
   const distPath = path.resolve('node_modules', 'magic-hook', 'dist')
 
-  await installPkgs(['magic-hook@2.0.0'], testDefaults())
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults())
 
   const distPathExists = await exists(distPath)
   t.ok(distPathExists, 'magic-hook@2.0.0 dist folder reinstalled')
@@ -331,7 +375,7 @@ test('refetch package to store if it has been modified', async function (t) {
 
 test("don't refetch package to store if it has been modified and verify-store-integrity = false", async (t: tape.Test) => {
   const project = prepare(t)
-  const opts = testDefaults({verifyStoreIntegrity: false})
+  const opts = await testDefaults({verifyStoreIntegrity: false})
   await installPkgs(['magic-hook@2.0.0'], opts)
 
   await writeJsonFile(path.join(await project.getStorePath(), 'localhost+4873', 'magic-hook', '2.0.0', 'node_modules', 'magic-hook', 'package.json'), {})
@@ -346,7 +390,7 @@ test("don't refetch package to store if it has been modified and verify-store-in
 // TODO: decide what to do with this case
 test['skip']('relink package to project if the dependency is not linked from store', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['magic-hook@2.0.0'], testDefaults({save: true, saveExact: true}))
+  await installPkgs(['magic-hook@2.0.0'], await testDefaults({save: true, saveExact: true}))
 
   const pkgJsonPath = path.resolve('node_modules', 'magic-hook', 'package.json')
 
@@ -363,14 +407,14 @@ test['skip']('relink package to project if the dependency is not linked from sto
 
   t.ok(storeInode !== await getInode(), 'package.json inode changed')
 
-  await install(testDefaults({repeatInstallDepth: 0}))
+  await install(await testDefaults({repeatInstallDepth: 0}))
 
   t.ok(storeInode === await getInode(), 'package.json inode matches the one that is in store')
 })
 
 test('circular deps', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['circular-deps-1-of-2'], testDefaults())
+  await installPkgs(['circular-deps-1-of-2'], await testDefaults())
 
   const m = project.requireModule('circular-deps-1-of-2/mirror')
 
@@ -379,15 +423,22 @@ test('circular deps', async function (t: tape.Test) {
   t.notOk(await exists(path.join('node_modules', 'circular-deps-1-of-2', 'node_modules', 'circular-deps-2-of-2', 'node_modules', 'circular-deps-1-of-2')), 'circular dependency is avoided')
 })
 
-test('concurrent circular deps', async function (t) {
+test('concurrent circular deps', async (t: tape.Test) => {
+  // es5-ext is an external package from the registry
+  // the latest dist-tag is overriden to have a stable test
+  await addDistTag('es5-ext', '0.10.31', 'latest')
+  await addDistTag('es6-iterator', '2.0.1', 'latest')
+
   const project = prepare(t)
-  await installPkgs(['es6-iterator@2.0.0'], testDefaults())
+  await installPkgs(['es6-iterator@2.0.0'], await testDefaults())
 
   const m = project.requireModule('es6-iterator')
 
   t.ok(m, 'es6-iterator is installed')
   t.ok(await exists(path.join('node_modules', '.localhost+4873', 'es6-iterator', '2.0.0', 'node_modules', 'es5-ext')))
   t.ok(await exists(path.join('node_modules', '.localhost+4873', 'es6-iterator', '2.0.1', 'node_modules', 'es5-ext')))
+  t.ok(await exists(path.join('node_modules', '.localhost+4873', 'es5-ext', '0.10.31', 'node_modules', 'es6-iterator')))
+  t.ok(await exists(path.join('node_modules', '.localhost+4873', 'es5-ext', '0.10.31', 'node_modules', 'es6-symbol')))
 })
 
 test('concurrent installation of the same packages', async function (t) {
@@ -395,7 +446,7 @@ test('concurrent installation of the same packages', async function (t) {
 
   // the same version of core-js is required by two different dependencies
   // of babek-core
-  await installPkgs(['babel-core@6.21.0'], testDefaults())
+  await installPkgs(['babel-core@6.21.0'], await testDefaults())
 
   const m = project.requireModule('babel-core')
 
@@ -404,17 +455,40 @@ test('concurrent installation of the same packages', async function (t) {
 
 test('big with dependencies and circular deps (babel-preset-2015)', async function (t) {
   const project = prepare(t)
-  await installPkgs(['babel-preset-es2015@6.3.13'], testDefaults())
+  await installPkgs(['babel-preset-es2015@6.3.13'], await testDefaults())
 
   const m = project.requireModule('babel-preset-es2015')
   t.ok(typeof m === 'object', 'babel-preset-es2015 is available')
 })
 
-test('bundleDependencies (pkg-with-bundled-dependencies@1.0.0)', async function (t) {
+test('bundledDependencies (pkg-with-bundled-dependencies@1.0.0)', async function (t: tape.Test) {
   const project = prepare(t)
-  await installPkgs(['pkg-with-bundled-dependencies@1.0.0'], testDefaults())
+
+  await installPkgs(['pkg-with-bundled-dependencies@1.0.0'], await testDefaults())
 
   await project.isExecutable('pkg-with-bundled-dependencies/node_modules/.bin/hello-world-js-bin')
+
+  const shr = await project.loadShrinkwrap()
+  t.deepEqual(
+    shr.packages['/pkg-with-bundled-dependencies/1.0.0'].bundledDependencies,
+    ['hello-world-js-bin'],
+    'bundledDependencies added to shrinkwrap.yaml'
+  )
+})
+
+test('bundleDependencies (pkg-with-bundle-dependencies@1.0.0)', async function (t: tape.Test) {
+  const project = prepare(t)
+
+  await installPkgs(['pkg-with-bundle-dependencies@1.0.0'], await testDefaults())
+
+  await project.isExecutable('pkg-with-bundle-dependencies/node_modules/.bin/hello-world-js-bin')
+
+  const shr = await project.loadShrinkwrap()
+  t.deepEqual(
+    shr.packages['/pkg-with-bundle-dependencies/1.0.0'].bundledDependencies,
+    ['hello-world-js-bin'],
+    'bundledDependencies added to shrinkwrap.yaml'
+  )
 })
 
 test('compiled modules (ursa@0.9.1)', async function (t) {
@@ -425,7 +499,7 @@ test('compiled modules (ursa@0.9.1)', async function (t) {
   }
 
   const project = prepare(t)
-  await installPkgs(['ursa@0.9.1'], testDefaults())
+  await installPkgs(['ursa@0.9.1'], await testDefaults())
 
   const m = project.requireModule('ursa')
   t.ok(typeof m === 'object', 'ursa() is available')
@@ -438,7 +512,7 @@ test('shrinkwrap compatibility', async function (t) {
   }
   const project = prepare(t, { dependencies: { rimraf: '*' } })
 
-  await installPkgs(['rimraf@2.5.1'], testDefaults())
+  await installPkgs(['rimraf@2.5.1'], await testDefaults())
 
   return new Promise((resolve, reject) => {
     const proc = crossSpawn.spawn('npm', ['shrinkwrap'])
@@ -455,55 +529,16 @@ test('shrinkwrap compatibility', async function (t) {
   })
 })
 
-test('production install (with --production flag)', async function (t) {
-  const project = prepare(t, basicPackageJson)
-
-  await install(testDefaults({ production: true }))
-
-  const rimrafDir = fs.statSync(path.resolve('node_modules', 'rimraf'))
-
-  let tapStatErrCode: number = 0
-  try {
-    fs.statSync(path.resolve('node_modules', '@rstacruz'))
-  } catch (err) {
-    tapStatErrCode = err.code
-  }
-
-  t.ok(rimrafDir.isSymbolicLink, 'rimraf exists')
-  t.is(tapStatErrCode, 'ENOENT', 'tap-spec does not exist')
-})
-
-test('production install (with production NODE_ENV)', async function (t) {
-  const originalNodeEnv = process.env.NODE_ENV
-  process.env.NODE_ENV = 'production'
-  const project = prepare(t, basicPackageJson)
-
-  await install(testDefaults())
-
-  // reset NODE_ENV
-  process.env.NODE_ENV = originalNodeEnv
-
-  const rimrafDir = fs.statSync(path.resolve('node_modules', 'rimraf'))
-
-  let tapStatErrCode: number = 0
-  try {
-    fs.statSync(path.resolve('node_modules', '@rstacruz'))
-  } catch (err) { tapStatErrCode = err.code }
-
-  t.ok(rimrafDir.isSymbolicLink, 'rimraf exists')
-  t.is(tapStatErrCode, 'ENOENT', 'tap-spec does not exist')
-})
-
 const wait = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))
 
 test('support installing into the same store simultaneously', async t => {
   const project = prepare(t)
   await Promise.all([
-    installPkgs(['pkg-that-installs-slowly'], testDefaults()),
+    installPkgs(['pkg-that-installs-slowly'], await testDefaults()),
     wait(500) // to be sure that lock was created
       .then(async () => {
         await project.storeHasNot('pkg-that-installs-slowly')
-        await installPkgs(['rimraf@2.5.1'], testDefaults())
+        await installPkgs(['rimraf@2.5.1'], await testDefaults())
       })
       .then(async  () => {
         await project.has('pkg-that-installs-slowly')
@@ -516,11 +551,11 @@ test('support installing into the same store simultaneously', async t => {
 test('support installing and uninstalling from the same store simultaneously', async t => {
   const project = prepare(t)
   await Promise.all([
-    installPkgs(['pkg-that-installs-slowly'], testDefaults()),
+    installPkgs(['pkg-that-installs-slowly'], await testDefaults()),
     wait(500) // to be sure that lock was created
       .then(async () => {
         await project.storeHasNot('pkg-that-installs-slowly')
-        await uninstall(['rimraf@2.5.1'], testDefaults())
+        await uninstall(['rimraf@2.5.1'], await testDefaults())
       })
       .then(async () => {
         await project.has('pkg-that-installs-slowly')
@@ -536,29 +571,28 @@ test('top-level packages should find the plugins they use', async function (t) {
       test: 'pkg-that-uses-plugins'
     }
   })
-  await installPkgs(['pkg-that-uses-plugins', 'plugin-example'], testDefaults({ save: true }))
+  await installPkgs(['pkg-that-uses-plugins', 'plugin-example'], await testDefaults({ save: true }))
   const result = spawnSync('npm', ['test'])
   t.ok(result.stdout.toString().indexOf('My plugin is plugin-example') !== -1, 'package executable have found its plugin')
   t.equal(result.status, 0, 'executable exited with success')
 })
 
-test('not top-level packages should find the plugins they use', async function (t) {
+test('not top-level packages should find the plugins they use', async function (t: tape.Test) {
   // standard depends on eslint and eslint plugins
   const project = prepare(t, {
     scripts: {
       test: 'standard'
     }
   })
-  await installPkgs(['standard@8.6.0'], testDefaults({ save: true }))
+  await installPkgs(['standard@8.6.0'], await testDefaults({ save: true }))
   const result = spawnSync('npm', ['test'])
-  console.log(result.stdout.toString())
   t.equal(result.status, 0, 'standard exited with success')
 })
 
 test('bin specified in the directories property linked to .bin folder', async function (t) {
   const project = prepare(t)
 
-  await installPkgs(['pkg-with-directories-bin'], testDefaults())
+  await installPkgs(['pkg-with-directories-bin'], await testDefaults())
 
   await project.isExecutable('.bin/pkg-with-directories-bin')
 })
@@ -569,7 +603,7 @@ test('run js bin file', async function (t) {
       test: 'hello-world-js-bin'
     }
   })
-  await installPkgs(['hello-world-js-bin'], testDefaults({ save: true }))
+  await installPkgs(['hello-world-js-bin'], await testDefaults({ save: true }))
 
   const result = spawnSync('npm', ['test'])
   t.ok(result.stdout.toString().indexOf('Hello world!') !== -1, 'package executable printed its message')
@@ -579,17 +613,17 @@ test('run js bin file', async function (t) {
 test('building native addons', async function (t) {
   const project = prepare(t)
 
-  await installPkgs(['runas@3.1.1'], testDefaults())
+  await installPkgs(['runas@3.1.1'], await testDefaults())
 
   t.ok(await exists(path.join('node_modules', 'runas', 'build')), 'build folder created')
 })
 
-test('should update subdep on second install', async function (t) {
+test('should update subdep on second install', async (t: tape.Test) => {
   const project = prepare(t)
 
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', 'latest')
 
-  await installPkgs(['pkg-with-1-dep'], testDefaults({save: true}))
+  await installPkgs(['pkg-with-1-dep'], await testDefaults({save: true}))
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.0.0')
 
@@ -599,7 +633,15 @@ test('should update subdep on second install', async function (t) {
 
   await addDistTag('dep-of-pkg-with-1-dep', '100.1.0', 'latest')
 
-  await install(testDefaults({depth: 1, update: true}))
+  const reporter = sinon.spy()
+
+  await install(await testDefaults({depth: 1, update: true, reporter}))
+
+  t.ok(reporter.calledWithMatch(<StatsLog>{
+    name: 'pnpm:stats',
+    level: 'debug',
+    added: 1,
+  }), 'added stat')
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.1.0')
 
@@ -616,7 +658,7 @@ test('should not update subdep when depth is smaller than depth of package', asy
 
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', 'latest')
 
-  await installPkgs(['pkg-with-1-dep'], testDefaults({save: true}))
+  await installPkgs(['pkg-with-1-dep'], await testDefaults({save: true}))
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.0.0')
 
@@ -626,7 +668,7 @@ test('should not update subdep when depth is smaller than depth of package', asy
 
   await addDistTag('dep-of-pkg-with-1-dep', '100.1.0', 'latest')
 
-  await install(testDefaults({depth: 0, update: true}))
+  await install(await testDefaults({depth: 0, update: true}))
 
   await project.storeHas('dep-of-pkg-with-1-dep', '100.0.0')
 
@@ -641,12 +683,12 @@ test('should not update subdep when depth is smaller than depth of package', asy
 test('should install dependency in second project', async function (t) {
   const project1 = prepare(t)
 
-  await installPkgs(['pkg-with-1-dep'], testDefaults({save: true, store: '../store'}))
+  await installPkgs(['pkg-with-1-dep'], await testDefaults({save: true, store: '../store'}))
   t.equal(project1.requireModule('pkg-with-1-dep')().name, 'dep-of-pkg-with-1-dep', 'can require in 1st pkg')
 
   const project2 = prepare(t)
 
-  await installPkgs(['pkg-with-1-dep'], testDefaults({save: true, store: '../store'}))
+  await installPkgs(['pkg-with-1-dep'], await testDefaults({save: true, store: '../store'}))
 
   t.equal(project2.requireModule('pkg-with-1-dep')().name, 'dep-of-pkg-with-1-dep', 'can require in 2nd pkg')
 })
@@ -654,10 +696,10 @@ test('should install dependency in second project', async function (t) {
 test('should throw error when trying to install using a different store then the previous one', async function(t) {
   const project = prepare(t)
 
-  await installPkgs(['rimraf@2.5.1'], testDefaults({store: 'node_modules/.store1'}))
+  await installPkgs(['rimraf@2.5.1'], await testDefaults({store: 'node_modules/.store1'}))
 
   try {
-    await installPkgs(['is-negative'], testDefaults({store: 'node_modules/.store2'}))
+    await installPkgs(['is-negative'], await testDefaults({store: 'node_modules/.store2'}))
     t.fail('installation should have failed')
   } catch (err) {
     t.equal(err.code, 'UNEXPECTED_STORE', 'failed with correct error code')
@@ -674,9 +716,10 @@ test('shrinkwrap locks npm dependencies', async function (t: tape.Test) {
 
   await addDistTag('dep-of-pkg-with-1-dep', '100.0.0', 'latest')
 
-  await installPkgs(['pkg-with-1-dep'], testDefaults({save: true, reporter}))
+  await installPkgs(['pkg-with-1-dep'], await testDefaults({save: true, reporter}))
 
   t.ok(reporter.calledWithMatch(<ProgressLog>{
+    name: 'pnpm:progress',
     level: 'debug',
     status: 'resolving_content',
     pkgId: 'localhost+4873/pkg-with-1-dep/100.0.0',
@@ -694,7 +737,7 @@ test('shrinkwrap locks npm dependencies', async function (t: tape.Test) {
   await rimraf('node_modules')
 
   reporter.reset()
-  await install(testDefaults({reporter}))
+  await install(await testDefaults({reporter}))
 
   t.ok(reporter.calledWithMatch(<ProgressLog>{
     level: 'debug',
@@ -715,7 +758,7 @@ test('shrinkwrap locks npm dependencies', async function (t: tape.Test) {
 test('self-require should work', async function (t) {
   const project = prepare(t)
 
-  await installPkgs(['uses-pkg-with-self-usage'], testDefaults())
+  await installPkgs(['uses-pkg-with-self-usage'], await testDefaults())
 
   t.ok(project.requireModule('uses-pkg-with-self-usage'))
 })
@@ -723,15 +766,15 @@ test('self-require should work', async function (t) {
 test('install on project with lockfile and no node_modules', async (t: tape.Test) => {
   const project = prepare(t)
 
-  await installPkgs(['is-negative'], testDefaults())
+  await installPkgs(['is-negative'], await testDefaults())
 
   await rimraf('node_modules')
 
-  await installPkgs(['is-positive'], testDefaults())
+  await installPkgs(['is-positive'], await testDefaults())
 
   t.ok(project.requireModule('is-positive'), 'installed new dependency')
 
-  t.ok(project.hasNot('is-negative'), 'did not reinstall removed dependency')
+  await project.hasNot('is-negative')
 })
 
 test('install a dependency with * range', async (t: tape.Test) => {
@@ -740,7 +783,21 @@ test('install a dependency with * range', async (t: tape.Test) => {
       'has-beta-only': '*',
     },
   })
-  await install(testDefaults())
+  const reporter = sinon.spy()
 
-  project.has('has-beta-only')
+  await install(await testDefaults({reporter}))
+
+  await project.has('has-beta-only')
+
+  t.ok(reporter.calledWithMatch(<PackageJsonLog>{
+    name: 'pnpm:package-json',
+    level: 'debug',
+    updated: {
+      name: 'project',
+      version: '0.0.0',
+      dependencies: {
+        'has-beta-only': '*',
+      },
+    },
+  }), 'should log package-json updated even when package.json was not changed')
 })
