@@ -4,7 +4,8 @@ import promisifyTape from 'tape-promise'
 import loadYamlFile = require('load-yaml-file')
 import exists = require('path-exists')
 import deepRequireCwd = require('deep-require-cwd')
-import {install, installPkgs} from '../../src'
+import sinon = require('sinon')
+import {install, installPkgs} from 'supi'
 import {
   prepare,
   addDistTag,
@@ -18,24 +19,26 @@ const test = promisifyTape(tape)
 test('successfully install optional dependency with subdependencies', async function (t) {
   const project = prepare(t)
 
-  await installPkgs(['fsevents@1.0.14'], testDefaults({saveOptional: true}))
+  await installPkgs(['fsevents@1.0.14'], await testDefaults({saveOptional: true}))
 })
 
-test('skip failing optional dependencies', async function (t) {
+test('skip failing optional dependencies', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['pkg-with-failing-optional-dependency@1.0.1'], testDefaults())
+  await installPkgs(['pkg-with-failing-optional-dependency@1.0.1'], await testDefaults())
 
   const m = project.requireModule('pkg-with-failing-optional-dependency')
   t.ok(m(-1), 'package with failed optional dependency has the dependencies installed correctly')
 })
 
-test('skip optional dependency that does not support the current OS', async function (t: tape.Test) {
+test('skip optional dependency that does not support the current OS', async (t: tape.Test) => {
   const project = prepare(t, {
     optionalDependencies: {
       'not-compatible-with-any-os': '*'
     }
   })
-  await install(testDefaults())
+  const reporter = sinon.spy()
+
+  await install(await testDefaults({reporter}))
 
   await project.hasNot('not-compatible-with-any-os')
   await project.storeHas('not-compatible-with-any-os', '1.0.0')
@@ -50,32 +53,55 @@ test('skip optional dependency that does not support the current OS', async func
     'localhost+4873/dep-of-optional-pkg/1.0.0',
     'localhost+4873/not-compatible-with-any-os/1.0.0',
   ])
+
+  const logMatcher = sinon.match({
+    level: 'warn',
+    message: 'Skipping failed optional dependency not-compatible-with-any-os@1.0.0',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+  t.equal(reportedTimes, 1, 'skipping optional dependency is logged')
 })
 
-test('skip optional dependency that does not support the current Node version', async function (t: tape.Test) {
+test('skip optional dependency that does not support the current Node version', async (t: tape.Test) => {
   const project = prepare(t, {
     optionalDependencies: {
       'for-legacy-node': '*'
     }
   })
+  const reporter = sinon.spy()
 
-  await install(testDefaults())
+  await install(await testDefaults({reporter}))
 
   await project.hasNot('for-legacy-node')
   await project.storeHas('for-legacy-node', '1.0.0')
+
+  const logMatcher = sinon.match({
+    level: 'warn',
+    message: 'Skipping failed optional dependency for-legacy-node@1.0.0',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+  t.equal(reportedTimes, 1, 'skipping optional dependency is logged')
 })
 
-test('skip optional dependency that does not support the current pnpm version', async function (t) {
+test('skip optional dependency that does not support the current pnpm version', async (t: tape.Test) => {
   const project = prepare(t, {
     optionalDependencies: {
       'for-legacy-pnpm': '*'
     }
   })
+  const reporter = sinon.spy()
 
-  await install(testDefaults())
+  await install(await testDefaults({reporter}))
 
   await project.hasNot('for-legacy-pnpm')
   await project.storeHas('for-legacy-pnpm', '1.0.0')
+
+  const logMatcher = sinon.match({
+    level: 'warn',
+    message: 'Skipping failed optional dependency for-legacy-pnpm@1.0.0',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+  t.equal(reportedTimes, 1, 'skipping optional dependency is logged')
 })
 
 test('don\'t skip optional dependency that does not support the current OS when forcing', async function (t) {
@@ -85,7 +111,7 @@ test('don\'t skip optional dependency that does not support the current OS when 
     }
   })
 
-  await install(testDefaults({
+  await install(await testDefaults({
     force: true
   }))
 
@@ -95,12 +121,20 @@ test('don\'t skip optional dependency that does not support the current OS when 
 
 test('optional subdependency is skipped', async (t: tape.Test) => {
   const project = prepare(t)
+  const reporter = sinon.spy()
 
-  await installPkgs(['pkg-with-optional', 'dep-of-optional-pkg'], testDefaults())
+  await installPkgs(['pkg-with-optional', 'dep-of-optional-pkg'], await testDefaults({reporter}))
 
   const modulesInfo = await loadYamlFile<{skipped: string[]}>(path.join('node_modules', '.modules.yaml'))
 
   t.deepEqual(modulesInfo.skipped, ['localhost+4873/not-compatible-with-any-os/1.0.0'], 'optional subdep skipped')
+
+  const logMatcher = sinon.match({
+    level: 'warn',
+    message: 'pkg-with-optional: Skipping failed optional dependency not-compatible-with-any-os@1.0.0',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+  t.equal(reportedTimes, 1, 'skipping optional dependency is logged')
 })
 
 test('not installing optional dependencies when optional is false', async (t: tape.Test) => {
@@ -113,10 +147,10 @@ test('not installing optional dependencies when optional is false', async (t: ta
     },
   })
 
-  await install(testDefaults({optional: false}))
+  await install(await testDefaults({optional: false}))
 
-  project.hasNot('is-positive')
-  project.has('pkg-with-good-optional')
+  await project.hasNot('is-positive')
+  await project.has('pkg-with-good-optional')
 
   t.ok(deepRequireCwd(['pkg-with-good-optional', 'dep-of-pkg-with-1-dep', './package.json']))
   t.notOk(deepRequireCwd.silent(['pkg-with-good-optional', 'is-positive', './package.json']), 'optional subdep not installed')
@@ -132,7 +166,7 @@ test('optional dependency has bigger priority than regular dependency', async (t
     },
   })
 
-  await install(testDefaults())
+  await install(await testDefaults())
 
   t.ok(deepRequireCwd(['is-positive', './package.json']).version, '3.1.0')
 })
