@@ -2,7 +2,7 @@ import tape = require('tape')
 import promisifyTape from 'tape-promise'
 import path = require('path')
 import exists = require('path-exists')
-import {installPkgs, install} from '../../src'
+import {installPkgs, install} from 'supi'
 import {
   prepare,
   testDefaults,
@@ -10,26 +10,32 @@ import {
 import deepRequireCwd = require('deep-require-cwd')
 import rimraf = require('rimraf-then')
 import sinon = require('sinon')
+import loadJsonFile = require('load-json-file')
 
 const test = promisifyTape(tape)
 const NM = 'node_modules'
 
-test("don't fail when peer dependency is fetched from GitHub", t => {
+test("don't fail when peer dependency is fetched from GitHub", async t => {
   const project = prepare(t)
-  return installPkgs(['test-pnpm-peer-deps'], testDefaults())
+  await installPkgs(['test-pnpm-peer-deps'], await testDefaults())
 })
 
 test('peer dependency is grouped with dependency when peer is resolved not from a top dependency', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['using-ajv'], testDefaults())
+  const opts = await testDefaults()
+  await installPkgs(['using-ajv'], opts)
 
   t.ok(await exists(path.join(NM, '.localhost+4873', 'ajv-keywords', '1.5.0', 'ajv@4.10.4', NM, 'ajv')), 'peer dependency is linked')
   t.equal(deepRequireCwd(['using-ajv', 'ajv-keywords', 'ajv', './package.json']).version, '4.10.4')
 
+  const storeIndex = await loadJsonFile(path.join(opts.store, 'store.json'))
+  t.ok(storeIndex['localhost+4873/ajv-keywords/1.5.0'], 'localhost+4873/ajv-keywords/1.5.0 added to store index')
+  t.ok(storeIndex['localhost+4873/using-ajv/1.0.0'], 'localhost+4873/using-ajv/1.0.0 added to store index')
+
   // testing that peers are reinstalled correctly using info from the shrinkwrap file
   await rimraf('node_modules')
   await rimraf(path.resolve('..', '.store'))
-  await install(testDefaults())
+  await install(await testDefaults())
 
   t.ok(await exists(path.join(NM, '.localhost+4873', 'ajv-keywords', '1.5.0', 'ajv@4.10.4', NM, 'ajv')), 'peer dependency is linked')
   t.equal(deepRequireCwd(['using-ajv', 'ajv-keywords', 'ajv', './package.json']).version, '4.10.4')
@@ -40,7 +46,7 @@ test('peer dependency is not grouped with dependent when the peer is a top depen
 
   const reporter = sinon.spy()
 
-  await installPkgs(['ajv@4.10.4', 'ajv-keywords@1.5.0'], testDefaults({reporter}))
+  await installPkgs(['ajv@4.10.4', 'ajv-keywords@1.5.0'], await testDefaults({reporter}))
 
   t.notOk(reporter.calledWithMatch({
     message: 'localhost+4873/ajv-keywords/1.5.0 requires a peer of ajv@>=4.10.0 but none was installed.',
@@ -49,24 +55,57 @@ test('peer dependency is not grouped with dependent when the peer is a top depen
   t.ok(await exists(path.join(NM, '.localhost+4873', 'ajv-keywords', '1.5.0', NM, 'ajv-keywords')), 'dependent is at the normal location')
 })
 
-test('warning is reported when cannot resolve peer dependency', async (t: tape.Test) => {
+test('warning is reported when cannot resolve peer dependency for top-level dependency', async (t: tape.Test) => {
   const project = prepare(t)
 
   const reporter = sinon.spy()
 
-  await installPkgs(['ajv-keywords@1.5.0'], testDefaults({reporter}))
+  await installPkgs(['ajv-keywords@1.5.0'], await testDefaults({reporter}))
 
-  t.ok(reporter.calledWithMatch({
-    message: 'localhost+4873/ajv-keywords/1.5.0 requires a peer of ajv@>=4.10.0 but none was installed.',
-  }), 'warning is logged about unresolved peer dep')
+  const logMatcher = sinon.match({
+    message: 'ajv-keywords@1.5.0 requires a peer of ajv@>=4.10.0 but none was installed.',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+
+  t.equal(reportedTimes, 1, 'warning is logged (once) about unresolved peer dep')
+})
+
+test('warning is reported when cannot resolve peer dependency for non-top-level dependency', async (t: tape.Test) => {
+  const project = prepare(t)
+
+  const reporter = sinon.spy()
+
+  await installPkgs(['abc-grand-parent-without-c'], await testDefaults({reporter}))
+
+  const logMatcher = sinon.match({
+    message: 'abc-grand-parent-without-c > abc-parent-with-ab: abc@1.0.0 requires a peer of peer-c@^1.0.0 but none was installed.',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+
+  t.equal(reportedTimes, 1, 'warning is logged (once) about unresolved peer dep')
+})
+
+test('warning is reported when bad version of resolved peer dependency for non-top-level dependency', async (t: tape.Test) => {
+  const project = prepare(t)
+
+  const reporter = sinon.spy()
+
+  await installPkgs(['abc-grand-parent-without-c', 'peer-c@2'], await testDefaults({reporter}))
+
+  const logMatcher = sinon.match({
+    message: 'abc-grand-parent-without-c > abc-parent-with-ab: abc@1.0.0 requires a peer of peer-c@^1.0.0 but version 2.0.0 was installed.',
+  })
+  const reportedTimes = reporter.withArgs(logMatcher).callCount
+
+  t.equal(reportedTimes, 1, 'warning is logged (once) about unresolved peer dep')
 })
 
 test('top peer dependency is not linked on subsequent install', async (t: tape.Test) => {
   const project = prepare(t)
 
-  await installPkgs(['ajv@4.10.4'], testDefaults())
+  await installPkgs(['ajv@4.10.4'], await testDefaults())
 
-  await installPkgs(['ajv-keywords@1.5.0'], testDefaults())
+  await installPkgs(['ajv-keywords@1.5.0'], await testDefaults())
 
   t.ok(await exists(path.join(NM, '.localhost+4873', 'ajv-keywords', '1.5.0', NM, 'ajv-keywords')), 'dependent is at the normal location')
   t.notOk(await exists(path.join(NM, '.localhost+4873', 'ajv-keywords', '1.5.0', 'ajv@4.10.4', NM, 'ajv')), 'peer dependency is not linked')
@@ -76,13 +115,15 @@ async function okFile (t: tape.Test, filename: string) {
   t.ok(await exists(filename), `exists ${filename}`)
 }
 
-test('peer dependencies are linked', async (t: tape.Test) => {
+// This usecase was failing. See https://github.com/pnpm/supi/issues/15
+test('peer dependencies are linked when running one named installation', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['abc-parent-with-ab', 'abc-grand-parent-with-c', 'peer-c@2.0.0'], testDefaults())
+
+  await installPkgs(['abc-grand-parent-with-c', 'abc-parent-with-ab', 'peer-c@2.0.0'], await testDefaults())
 
   const pkgVariationsDir = path.join(NM, '.localhost+4873', 'abc', '1.0.0')
 
-  const pkgVariation1 = path.join(pkgVariationsDir, 'peer-a@1.0.0+peer-b@1.0.0+peer-c@1.0.0', NM)
+  const pkgVariation1 = path.join(pkgVariationsDir, '165e1e08a3f7e7f77ddb572ad0e55660', NM)
   await okFile(t, path.join(pkgVariation1, 'abc'))
   await okFile(t, path.join(pkgVariation1, 'peer-a'))
   await okFile(t, path.join(pkgVariation1, 'peer-b'))
@@ -99,9 +140,67 @@ test('peer dependencies are linked', async (t: tape.Test) => {
   t.equal(deepRequireCwd(['abc-grand-parent-with-c', 'abc-parent-with-ab', 'abc', 'peer-c', './package.json']).version, '1.0.0')
 })
 
+test('peer dependencies are linked when running two separate named installations', async (t: tape.Test) => {
+  const project = prepare(t)
+  await installPkgs(['abc-grand-parent-with-c', 'peer-c@2.0.0'], await testDefaults())
+  await installPkgs(['abc-parent-with-ab'], await testDefaults())
+
+  const pkgVariationsDir = path.join(NM, '.localhost+4873', 'abc', '1.0.0')
+
+  const pkgVariation1 = path.join(pkgVariationsDir, '165e1e08a3f7e7f77ddb572ad0e55660', NM)
+  await okFile(t, path.join(pkgVariation1, 'abc'))
+  await okFile(t, path.join(pkgVariation1, 'peer-a'))
+  await okFile(t, path.join(pkgVariation1, 'peer-b'))
+  await okFile(t, path.join(pkgVariation1, 'peer-c'))
+  await okFile(t, path.join(pkgVariation1, 'dep-of-pkg-with-1-dep'))
+
+  const pkgVariation2 = path.join(pkgVariationsDir, 'peer-a@1.0.0+peer-b@1.0.0', NM)
+  await okFile(t, path.join(pkgVariation2, 'abc'))
+  await okFile(t, path.join(pkgVariation2, 'peer-a'))
+  await okFile(t, path.join(pkgVariation2, 'peer-b'))
+  await okFile(t, path.join(pkgVariation2, 'dep-of-pkg-with-1-dep'))
+
+  t.equal(deepRequireCwd(['abc-parent-with-ab', 'abc', 'peer-c', './package.json']).version, '2.0.0')
+  t.equal(deepRequireCwd(['abc-grand-parent-with-c', 'abc-parent-with-ab', 'abc', 'peer-c', './package.json']).version, '1.0.0')
+})
+
+test['skip']('peer dependencies are linked', async (t: tape.Test) => {
+  const project = prepare(t, {
+    dependencies: {
+      'abc-grand-parent-with-c': '*',
+      'peer-c': '2.0.0',
+    },
+    devDependencies: {
+      'abc-parent-with-ab': '*',
+    },
+  })
+  await install(await testDefaults())
+
+  const pkgVariationsDir = path.join(NM, '.localhost+4873', 'abc', '1.0.0')
+
+  const pkgVariation1 = path.join(pkgVariationsDir, '165e1e08a3f7e7f77ddb572ad0e55660', NM)
+  await okFile(t, path.join(pkgVariation1, 'abc'))
+  await okFile(t, path.join(pkgVariation1, 'peer-a'))
+  await okFile(t, path.join(pkgVariation1, 'peer-b'))
+  await okFile(t, path.join(pkgVariation1, 'peer-c'))
+  await okFile(t, path.join(pkgVariation1, 'dep-of-pkg-with-1-dep'))
+
+  const pkgVariation2 = path.join(pkgVariationsDir, 'peer-a@1.0.0+peer-b@1.0.0', NM)
+  await okFile(t, path.join(pkgVariation2, 'abc'))
+  await okFile(t, path.join(pkgVariation2, 'peer-a'))
+  await okFile(t, path.join(pkgVariation2, 'peer-b'))
+  await okFile(t, path.join(pkgVariation2, 'dep-of-pkg-with-1-dep'))
+
+  t.equal(deepRequireCwd(['abc-parent-with-ab', 'abc', 'peer-c', './package.json']).version, '2.0.0')
+  t.equal(deepRequireCwd(['abc-grand-parent-with-c', 'abc-parent-with-ab', 'abc', 'peer-c', './package.json']).version, '1.0.0')
+
+  const shr = await project.loadShrinkwrap()
+  t.ok(shr.packages['/abc-parent-with-ab/1.0.0/peer-a@1.0.0+peer-b@1.0.0'].dev, 'the dev resolution set is marked as dev in shrinkwrap.yaml')
+})
+
 test('scoped peer dependency is linked', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['for-testing-scoped-peers'], testDefaults())
+  await installPkgs(['for-testing-scoped-peers'], await testDefaults())
 
   const pkgVariation = path.join(NM, '.localhost+4873', '@having', 'scoped-peer', '1.0.0', '@scoped!peer@1.0.0', NM)
   await okFile(t, path.join(pkgVariation, '@having', 'scoped-peer'))
@@ -111,7 +210,7 @@ test('scoped peer dependency is linked', async (t: tape.Test) => {
 test('peer bins are linked', async (t: tape.Test) => {
   const project = prepare(t)
 
-  await installPkgs(['for-testing-peers-having-bins'], testDefaults())
+  await installPkgs(['for-testing-peers-having-bins'], await testDefaults())
 
   const pkgVariation = path.join('.localhost+4873', 'pkg-with-peer-having-bin', '1.0.0', 'peer-with-bin@1.0.0', NM)
 
@@ -122,7 +221,7 @@ test('peer bins are linked', async (t: tape.Test) => {
 
 test('run pre/postinstall scripts of each variations of packages with peer dependencies', async (t: tape.Test) => {
   const project = prepare(t)
-  await installPkgs(['parent-of-pkg-with-events-and-peers', 'pkg-with-events-and-peers', 'peer-c@2.0.0'], testDefaults())
+  await installPkgs(['parent-of-pkg-with-events-and-peers', 'pkg-with-events-and-peers', 'peer-c@2.0.0'], await testDefaults())
 
   const pkgVariation1 = path.join(NM, '.localhost+4873', 'pkg-with-events-and-peers', '1.0.0', 'peer-c@1.0.0', NM)
   await okFile(t, path.join(pkgVariation1, 'pkg-with-events-and-peers', 'generated-by-preinstall.js'))
@@ -139,17 +238,23 @@ test('package that resolves its own peer dependency', async (t: tape.Test) => {
   // does it currently print a warning that peer dependency is not resolved?
 
   const project = prepare(t)
-  await installPkgs(['pkg-with-resolved-peer', 'peer-c@2.0.0'], testDefaults())
+  await installPkgs(['pkg-with-resolved-peer', 'peer-c@2.0.0'], await testDefaults())
 
   t.equal(deepRequireCwd(['pkg-with-resolved-peer', 'peer-c', './package.json']).version, '1.0.0')
 
   t.ok(await exists(path.join(NM, '.localhost+4873', 'pkg-with-resolved-peer', '1.0.0', NM, 'pkg-with-resolved-peer')))
+
+  const shr = await project.loadShrinkwrap()
+
+  t.notOk(shr.packages['/pkg-with-resolved-peer/1.0.0'].peerDependencies, 'peerDependencies not added to shrinkwrap')
+  t.ok(shr.packages['/pkg-with-resolved-peer/1.0.0'].dependencies['peer-c'])
+  t.ok(shr.packages['/pkg-with-resolved-peer/1.0.0'].optionalDependencies['peer-b'])
 })
 
 test('own peer installed in root as well is linked to root', async function (t: tape.Test) {
   const project = prepare(t)
 
-  await installPkgs(['is-negative@kevva/is-negative#2.1.0', 'peer-deps-in-child-pkg'], testDefaults())
+  await installPkgs(['is-negative@kevva/is-negative#2.1.0', 'peer-deps-in-child-pkg'], await testDefaults())
 
   t.ok(deepRequireCwd.silent(['is-negative', './package.json']), 'is-negative is linked to root')
 })
